Refactor hot code reloading
This is a major refactoring of hot code reloading. It now behaves much
better, finds the origin of the exceptions much deeper in the traceback,
but only reloads within the scope of the module. It handles class
instantiations, class methods, lambdas, and soon module-level functions.
There's no limit to how deep the exception is in the call stack, as long
as it can be traced back to the module's underlying function.